+++ XboxHomebrewStore/Font.h
@@ -13,7 +13,7 @@ typedef struct
     float    offset;
     uint32_t lastTick;
     float    pauseTimer;
-    int      textWidth;
+    float    textWidth;
     bool     active;
 } ScrollState;
 
@@ -21,8 +21,8 @@ class Font
 {
 public:
     static void Init();
-    static int MeasureText(const FontType font, const std::string& message);
-    static std::string TruncateText(const FontType font, const std::string& message, int maxWidth);
-    static void DrawText(const FontType font, const std::string message, uint32_t color, int x, int y);
-    static void DrawTextScrolling(const FontType font, const std::string& message, uint32_t color, int x, int y, int maxWidth, ScrollState* scrollState);
+    static float MeasureText(const FontType font, const std::string& message);
+    static std::string TruncateText(const FontType font, const std::string& message, float maxWidth);
+    static void DrawText(const FontType font, const std::string message, uint32_t color, float x, float y);
+    static void DrawTextScrolling(const FontType font, const std::string& message, uint32_t color, float x, float y, float maxWidth, ScrollState* scrollState);
 };
